requirements.txt
azure-identity>=1.20,<2.0
azure-search-documents>=11.5,<12.0
fastapi>=0.115,<1.0
openai>=1.65,<2.0